requirements.txt
schwab-py
python-dotenv
pyyaml
fastapi
uvicorn[standard]
plyer
httpx
pytest
pytz
